+++ helpers/note.js
@@ -24,7 +24,6 @@ import { getFolderListMinusExclusions, getFolderFromFilename } from '@helpers/fo
 import { displayTitle, type headingLevelType } from '@helpers/general'
 import { toNPLocaleDateString } from '@helpers/NPdateTime'
 import { findEndOfActivePartOfNote, findStartOfActivePartOfNote } from '@helpers/paragraph'
-import { formRegExForUsersOpenTasks } from '@helpers/regex'
 import { sortListBy } from '@helpers/sorting'
 import {
   isOpen,
