+++ np.Preview/src/previewMain.js
@@ -5,26 +5,18 @@
 // by Jonathan Clark, last updated 11.8.2023 for v0.4.?
 //--------------------------------------------------------------
 
-import pluginJson from '../plugin.json'
+
 // import open, { openApp, apps } from 'open'
-import showdown from 'showdown' // for Markdown -> HTML from https://github.com/showdownjs/showdown
-import { getCodeBlocksOfType } from '@helpers/codeBlocks'
-import { clo, JSP, logDebug, logError, logInfo, logWarn } from '@helpers/dev'
+import pluginJson from '../plugin.json'
+import { logDebug, logError, logWarn } from '@helpers/dev'
 import { addTrigger } from '@helpers/NPFrontMatter'
 import { displayTitle } from '@helpers/general'
+
 import {
-  getFrontMatterParagraphs,
-  hasFrontMatter
-} from '@helpers/NPFrontMatter'
-import {
-  convertHashtagsToHTML,
-  convertHighlightsToHTML,
-  convertMentionsToHTML,
-  convertUnderlinedToHTML,
+  getNoteContentAsHTML,
   type HtmlWindowOptions,
   showHTMLV2
 } from '@helpers/HTMLView'
-import { formRegExForUsersOpenTasks, RE_SYNC_MARKER } from '@helpers/regex'
 import { showMessageYesNo } from '@helpers/userInput'
 
 //--------------------------------------------------------------
@@ -82,40 +74,10 @@ Button a { text-decoration: none; font-size: 0.9rem; }
  * @author @jgclark
  * @param {string?} mermaidTheme name (optional)
  */
-export function previewNote(mermaidTheme?: string): void {
+export async function previewNote(mermaidTheme?: string): void {
   try {
     const { note, content, title } = Editor
-    let lines = content?.split('\n') ?? []
-    let hasFrontmatter = hasFrontMatter(content ?? '')
-    const RE_OPEN_TASK_FOR_USER = formRegExForUsersOpenTasks(false)
-
-    // Work on a copy of the note's content
-    // Change frontmatter for this note (if present)
-    // In particular remove trigger line
-    if (hasFrontmatter) {
-      let titleAsMD = ''
-      lines = lines.filter(l => l !== 'triggers: onEditorWillSave => np.Preview.updatePreview')
-      // look for 2nd '---' and double it, because of showdown bug
-      for (let i = 1; i < lines.length; i++) {
-        if (lines[i].match(/^title:\s/)) {
-          titleAsMD = lines[i].replace('title:', '#')
-          logDebug('previewNote', `removing title line ${String(i)}`)
-          lines.splice(i, 1)
-        }
-        if (lines[i].trim() === '---') {
-          lines.splice(i, 0, '') // add a blank before second HR to stop it acting as an ATX header line
-          lines.splice(i + 2, 0, titleAsMD) // add the title (as MD)
-          break
-        }
-      }
-
-      // If we now have empty frontmatter (so, just 3 sets of '---'), then remove them all
-      if (lines[0] === '---' && lines[1] === '' && lines[2] === '---') {
-        lines.splice(0, 3)
-        hasFrontmatter = false
-      }
-    }
-
+    const lines = content?.split('\n') ?? []
 
     // Update mermaid fenced code blocks to suitable <divs>
     // Note: did try to use getCodeBlocksOfType() helper but found it wasn't architected helpfully for this use case
@@ -133,89 +95,18 @@ export function previewNote(mermaidTheme?: string): void {
       }
     }
 
-    // Make some necessary changes before conversion to HTML
-    for (let i = 0; i < lines.length; i++) {
-      // remove any sync link markers (blockIds)
-      lines[i] = lines[i].replace(/\^[A-z0-9]{6}([^A-z0-9]|$)/g, '').trimRight()
+    let body = await getNoteContentAsHTML(lines.join('\n'), note)
 
-      // change open tasks to GFM-flavoured task syntax
-      const res = lines[i].match(RE_OPEN_TASK_FOR_USER)
-      if (res) {
-        lines[i] = lines[i].replace(res[0], '- [ ]')
-      }
-    }
-
-    // Make this proper Markdown -> HTML via showdown library
-    // Set some options to turn on various more advanced HTML conversions (see actual code at https://github.com/showdownjs/showdown/blob/master/src/options.js#L109):
-    const converterOptions = {
-      emoji: true,
-      footnotes: true,
-      ghCodeBlocks: true,
-      strikethrough: true,
-      tables: true,
-      tasklists: true,
-      metadata: false, // otherwise metadata is swallowed
-      requireSpaceBeforeHeadingText: true,
-      simpleLineBreaks: true // Makes this GFM style. TODO: make an option?
-    }
-    const converter = new showdown.Converter(converterOptions)
-    let body = converter.makeHtml(lines.join(`\n`))
-
-    // logDebug(pluginJson, 'Converter produces:\n' + body)
-
-    // TODO: Ideally build a frontmatter styler extension (to use above) but for now ...
-    // Tweak body output to put frontmatter in a box if it exists
-    if (hasFrontmatter) {
-      // replace first '<hr />' with start of div
-      body = body.replace('<hr />', '<div class="frontmatter">')
-      // replace what is now the first '<hr />' with end of div
-      body = body.replace('<hr />', '</div>')
-    }
-    // logDebug(pluginJson, body)
-
-    // Make other changes to the HTML to cater for NotePlan-specific syntax
-    lines = body.split('\n')
-    let modifiedLines = []
-    for (let line of lines) {
-      const origLine = line
-
-      // Display hashtags with .hashtag style
-      line = convertHashtagsToHTML(line)
-
-      // Display mentions with .attag style
-      line = convertMentionsToHTML(line)
-
-      // Display highlights with .highlight style
-      line = convertHighlightsToHTML(line)
-
-      // Replace [[notelinks]] with just underlined notelink
-      let captures = line.match(/\[\[(.*?)\]\]/)
-      if (captures) {
-        // clo(captures, 'results from [[notelinks]] match:')
-        for (let capturedTitle of captures) {
-          line = line.replace('[[' + capturedTitle + ']]', '~' + capturedTitle + '~')
-        }
-      }
-      // Display underlining with .underlined style
-      line = convertUnderlinedToHTML(line)
-
-      // Remove any blockIDs
-      line = line.replace(RE_SYNC_MARKER, '')
-
-      if (line !== origLine) {
-        logDebug('previewNote', `modified {${origLine}} -> {${line}}`)
-      }
-      modifiedLines.push(line)
-    }
     // Add mermaid script if needed
-    const finalBody = modifiedLines.join('\n') + (includesMermaid ? initMermaidScripts(mermaidTheme) : '')
-    console.log(initMermaidScripts("green"))
-
+    if (includesMermaid) {
+      body = initMermaidScripts(mermaidTheme) + body
+    }
     // Add sticky button at top right offering to print
     // (But printing doesn't work on i(Pad)OS ...)
     if (NotePlan.environment.platform === 'macOS') {
-      body = `	<div class="stickyButton"><button class="nonPrinting" type="printButton"><a href="preview.html" onclick="window.open(this.href).print(); return false;">Print (opens in system browser)</a></button></div>\n` + body // Note: seems to need the .print() even though it doesn't activate in the browser.
+      body = `<div class="stickyButton"><button class="nonPrinting" type="printButton"><a href="preview.html" onclick="window.open(this.href).print(); return false;">Print (opens in system browser)</a></button></div>\n${body}` // Note: seems to need the .print() even though it doesn't activate in the browser.
     }
+
     const headerTags = `<meta name="generator" content="np.Preview plugin by @jgclark v${pluginJson['plugin.version'] ?? '?'}">
 <meta name="date" content="${new Date().toISOString()}">`
 
@@ -234,7 +125,7 @@ export function previewNote(mermaidTheme?: string): void {
       shouldFocus: false, // shouuld not focus, if Window already exists
       // not setting defaults for x, y, width, height
     }
-    showHTMLV2(finalBody, windowOpts)
+    showHTMLV2(body, windowOpts)
     // logDebug('preview', `written results to HTML`)
   }
   catch (error) {
@@ -249,7 +140,7 @@ export function previewNote(mermaidTheme?: string): void {
 export async function openPreviewNoteInBrowser(): Promise<void> {
   try {
     // Call preview note function with 'default' theme (best for printing)
-    previewNote('default')
+    await previewNote('default')
     logDebug(pluginJson, `openPreviewNoteInBrowser: preview created; now will try to open in browser`)
     // FIXME: the following doesn't work -- something to do with imports and builtins
     // await open(savedFilename)
@@ -278,7 +169,7 @@ export async function addTriggerAndStartPreview(): Promise<void> {
     }
 
     // Start the preview
-    previewNote()
+    await previewNote()
   }
   catch (error) {
     logError(pluginJson, `${error.name}: ${error.message}`)
